+++ triple_arena/CHANGELOG.md
@@ -7,6 +7,10 @@
 - Added generation counter checking to all fuzz tests
 - Added several special cases involving single link chains for `ChainArena` tests
 
+### Additions
+- Added `SurjectArena`
+- Added `ChainArena::remove_chain`
+
 ## [0.7.0] - 2022-12-07
 ### Additions
 - added `first_ptr` and `next_ptr` functions